feat: build and publish ex_doc on merge (#5)
This completes the "Published documentation to GitHub pages" feature.

On CI, we build the docs to ensure there are no issues. On merge to
`main`, we build and publish them to GitHub pages. Usually I'd merge and
build docs on an actual release, but those will be somewhat rare here.