feat: setup benchmarking with benchee (#35)
This adds benchmarking in the `benchmarks` folder. I used the same
layout as the `test` folder with a single file entry that wraps up all
of the other files for organization. I also added a support module that
saves and loads results. This allows CI to cache the benchmark results
and display them for all CI runs. Great for checking performance
sensitive code in PRs.